UI - Refactoring workspace, admin unit and unit pages; UI - New summary tab in ws, admin unit and unit pages; Server - Adjusting API for landing page of workspace, admin unit and units; Server - New summary API; Server - Search cases service improvements;

+++ src/main/java/org/msh/etbm/commons/entities/query/QueryBuilderImpl.java
@@ -424,7 +424,7 @@ public class QueryBuilderImpl<E> implements QueryBuilder<E> {
                 res.setCount(getCount());
             } else {
                 // if there is no paging, avoid unnecessary database query
-                res.setCount(res.getList().size());
+                res.setCount((long)res.getList().size());
             }
 
             return res;
